Extracted Testers and Label Util into own files
