+++ src/components/renderers/controls/number/number-directive.ts
@@ -1,5 +1,5 @@
-
-import {AbstractControl,  schemaTypeIs} from '../abstract-control';
+import {AbstractControl} from '../abstract-control';
+import {schemaTypeIs} from "../../testers";
 
 class NumberDirective implements ng.IDirective {
     restrict = 'E';
