+++ src/components/renderers/layouts/categories/categorization-directive.ts
@@ -1,5 +1,5 @@
 import {AbstractLayout} from '../abstract-layout';
-import {uiTypeIs} from '../../controls/abstract-control';
+import {uiTypeIs} from "../../testers";
 
 class CategorizationDirective implements ng.IDirective {
     restrict = 'E';
