+++ src/components/renderers/layouts/masterdetail/masterdetail-directives.ts
@@ -1,6 +1,7 @@
-import {AbstractControl, uiTypeIs} from '../../controls/abstract-control';
+import {AbstractControl} from '../../controls/abstract-control';
 import {SchemaElement} from '../../../../jsonschema';
 import {PathResolver} from '../../../services/pathresolver/jsonforms-pathresolver';
+import {uiTypeIs} from "../../testers";
 
 class MasterDetailDirective implements ng.IDirective {
     restrict = 'E';
