+++ src/components/renderers/layouts/vertical/vertical-directive.ts
@@ -1,5 +1,5 @@
 import {AbstractLayout} from '../abstract-layout';
-import {uiTypeIs} from '../../controls/abstract-control';
+import {uiTypeIs} from "../../testers";
 
 class VerticalDirective implements ng.IDirective {
     restrict = 'E';
